requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via web3
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via httpx
apscheduler==3.10.4
    # via python-telegram-bot
attrs==24.2.0
    # via aiohttp
bitarray==2.9.2
    # via eth-account
build==1.2.1
    # via pip-tools
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
ckzg==2.0.0
    # via eth-account
click==8.1.7
    # via pip-tools
cytoolz==0.12.3
    # via eth-utils
eth-abi==5.1.0
    # via
    #   eth-account
    #   web3
eth-account==0.13.3
    # via web3
eth-hash[pycryptodome]==0.7.0
    # via
    #   eth-utils
    #   web3
eth-keyfile==0.8.1
    # via eth-account
eth-keys==0.5.1
    # via
    #   eth-account
    #   eth-keyfile
eth-rlp==2.1.0
    # via eth-account
eth-typing==5.0.0
    # via
    #   eth-abi
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==5.0.0
    # via
    #   eth-abi
    #   eth-account
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
hexbytes==1.2.1
    # via
    #   eth-account
    #   eth-rlp
    #   eth-utils
    #   web3
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via python-telegram-bot
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
packaging==24.1
    # via build
parsimonious==0.10.0
    # via eth-abi
pip-tools==7.4.1
    # via -r requirements.in
pycryptodome==3.20.0
    # via
    #   eth-hash
    #   eth-keyfile
pydantic==2.8.2
    # via
    #   eth-account
    #   web3
pydantic-core==2.20.1
    # via pydantic
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via -r requirements.in
python-telegram-bot[job-queue]==21.5
    # via -r requirements.in
pytz==2024.1
    # via
    #   apscheduler
    #   python-telegram-bot
pyunormalize==15.1.0
    # via web3
regex==2024.7.24
    # via parsimonious
requests==2.32.3
    # via
    #   -r requirements.in
    #   web3
rlp==4.0.1
    # via
    #   eth-account
    #   eth-rlp
six==1.16.0
    # via apscheduler
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
tenacity==9.0.0
    # via -r requirements.in
toolz==0.12.1
    # via cytoolz
types-requests==2.32.0.20240712
    # via web3
typing-extensions==4.12.2
    # via
    #   eth-typing
    #   pydantic
    #   pydantic-core
    #   web3
tzlocal==5.2
    # via apscheduler
urllib3==2.2.2
    # via
    #   requests
    #   types-requests
web3==7.2.0
    # via -r requirements.in
websockets==13.0.1
    # via web3
wheel==0.44.0
    # via pip-tools
yarl==1.9.7
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
